Refactor: put in own files ProfileResultsTest.kt and CategoryContents.kt
