+++ app/src/main/java/com/arygm/quickfix/ui/search/SearchOnBoarding.kt
@@ -8,12 +8,8 @@ import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyListState
-import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
@@ -23,9 +19,7 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme.colorScheme
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateListOf
@@ -34,23 +28,16 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.unit.dp
-import com.arygm.quickfix.MainActivity
-import com.arygm.quickfix.R
-import com.arygm.quickfix.model.account.Account
 import com.arygm.quickfix.model.account.AccountViewModel
-import com.arygm.quickfix.model.category.Category
 import com.arygm.quickfix.model.category.CategoryViewModel
-import com.arygm.quickfix.model.profile.WorkerProfile
 import com.arygm.quickfix.model.search.SearchViewModel
 import com.arygm.quickfix.ui.elements.QuickFixButton
 import com.arygm.quickfix.ui.elements.QuickFixTextFieldCustom
 import com.arygm.quickfix.ui.navigation.NavigationActions
 import com.arygm.quickfix.ui.navigation.TopLevelDestinations
 import com.arygm.quickfix.ui.theme.poppinsTypography
-import com.arygm.quickfix.utils.LocationHelper
 
 @Composable
 fun SearchOnBoarding(
@@ -152,7 +139,7 @@ fun SearchOnBoarding(
                   )
                 } else {
                   // Show Profiles
-                  ProfileContent(
+                  ProfileResults(
                       profiles = profiles,
                       searchViewModel = searchViewModel,
                       accountViewModel = accountViewModel,
@@ -164,116 +151,4 @@ fun SearchOnBoarding(
               }
         })
   }
-}
-
-@Composable
-fun CategoryContent(
-    navigationActions: NavigationActions,
-    searchViewModel: SearchViewModel,
-    listState: LazyListState,
-    expandedStates: MutableList<Boolean>,
-    itemCategories: List<Category>,
-    widthRatio: Float,
-    heightRatio: Float,
-) {
-  Column(
-      modifier = Modifier.fillMaxWidth().padding(horizontal = 10.dp * widthRatio),
-      horizontalAlignment = Alignment.Start) {
-        Text(
-            text = "Categories",
-            style = poppinsTypography.labelLarge,
-            color = colorScheme.onBackground,
-        )
-        Spacer(modifier = Modifier.height(4.dp * heightRatio))
-        LazyColumn(modifier = Modifier.fillMaxWidth(), state = listState) {
-          itemsIndexed(itemCategories, key = { index, _ -> index }) { index, item ->
-            ExpandableCategoryItem(
-                item = item,
-                isExpanded = expandedStates[index],
-                onExpandedChange = { expandedStates[index] = it },
-                searchViewModel = searchViewModel,
-                navigationActions = navigationActions,
-            )
-            Spacer(modifier = Modifier.height(10.dp * heightRatio))
-          }
-        }
-      }
-}
-
-@Composable
-fun ProfileContent(
-    profiles: List<WorkerProfile>, // Assuming WorkerProfile is your model class
-    listState: LazyListState,
-    searchViewModel: SearchViewModel,
-    accountViewModel: AccountViewModel,
-    widthRatio: Float,
-    heightRatio: Float,
-) {
-  // Column for wrapping the list
-  Column(
-      modifier = Modifier.fillMaxWidth().padding(horizontal = 10.dp * widthRatio),
-      horizontalAlignment = Alignment.Start) {
-        // Title: "Profiles"
-        Text(
-            text = "Profiles",
-            style = poppinsTypography.labelLarge,
-            color = colorScheme.onBackground,
-        )
-        Spacer(modifier = Modifier.height(4.dp * heightRatio))
-
-        // LazyColumn for displaying profiles
-        LazyColumn(modifier = Modifier.fillMaxWidth(), state = listState) {
-          items(profiles.size) { index ->
-            val profile = profiles[index]
-            var account by remember { mutableStateOf<Account?>(null) }
-            var distance by remember { mutableStateOf<Int?>(null) }
-
-            // Get user's current location and calculate distance
-            val locationHelper = LocationHelper(LocalContext.current, MainActivity())
-            locationHelper.getCurrentLocation { location ->
-              location?.let {
-                distance =
-                    profile.location?.let { workerLocation ->
-                      searchViewModel
-                          .calculateDistance(
-                              workerLocation.latitude,
-                              workerLocation.longitude,
-                              it.latitude,
-                              it.longitude)
-                          .toInt()
-                    }
-              }
-            }
-
-            // Fetch user account details
-            LaunchedEffect(profile.uid) {
-              accountViewModel.fetchUserAccount(profile.uid) { fetchedAccount ->
-                account = fetchedAccount
-              }
-            }
-
-            // Render profile card if account data is available
-            account?.let { acc ->
-              SearchWorkerProfileResult(
-                  modifier =
-                      Modifier.padding(vertical = 10.dp * heightRatio)
-                          .fillMaxWidth()
-                          .testTag("worker_profile_result_$index"),
-                  profileImage = R.drawable.placeholder_worker, // Replace with actual image
-                  name = "${acc.firstName} ${acc.lastName}",
-                  category = profile.fieldOfWork,
-                  rating = profile.rating,
-                  reviewCount = profile.reviews.size,
-                  location = profile.location?.name ?: "Unknown",
-                  price = profile.price.toString(),
-                  distance = distance,
-                  onBookClick = {
-                    // Handle book action
-                  })
-            }
-
-            Spacer(modifier = Modifier.height(10.dp * heightRatio))
-          }
-        }
-      }
 }
